Refactor artifact fetching into a class
